Update structs.livemd
Changed Hero struct key :hero_name to :name because on line 173 `"#{greeting} I am #{hero.name}"` this is to be outputed. So accessing hero.name raises key not found error.
And interchanged the arity of introduced and reveal function. On line 166 * create a `reveal/1` function which takes in an instance of a hero and returns 
and line 170 * create an `introduce/2` function which takes in an instance of a hero and a greeting
arity of functions to be defined are different.