Grammar and structure cleanup
Removed duplicate link to Ecto docs. This link was made one sentence prior.  
Small grammar clean up.  

The following sentence appears more than once in the reading. It is relevant in both places I noticed it, however it has not been the norm to so closely repeat data in a reading. 
> It's important to note that Ecto will only apply unapplied migrations, so you can safely run the ecto.migrate task multiple times without worrying about making the same changes to the database more than once. This can be very useful for managing the evolution of your database over time.