Clean up escaping string interpolation example
The whole special character being escaped is the string interpolation`#{}`; you can thing of this as escaping the leading hash symbol leaves everything else as just normal characters. It is unnecessary to escape each individual character in the string interpolation.  
You wouldn't need to do `\#\{\2\+\2\}` to include `#{2+2}` in a string rather that have Elixir interpolate it and return 4, `\#{2+2}` will do that. The whole interpolation is escaped with a single backslash.